requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
Flask-Cors==4.0.2
python-dotenv==1.0.1
requests==2.32.3
marshmallow==3.21.2
